requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
python-multipart==0.0.6
supabase==1.0.3
PyJWT==2.8.0
openai==1.3.0
PyMuPDF==1.23.5
pdfplumber==0.10.2
PyPDF2==3.0.1
pytesseract==0.3.10
google-api-python-client==2.107.0
opencv-python==4.8.1.78
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
playwright==1.40.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1